backend/requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
diskcache
fastapi
h11
idna
Jinja2
llama_cpp_python
MarkupSafe
numpy
pydantic
requests
sniffio
starlette
toml
uvicorn
urllib3
typing_extensions
typing-inspection
